\name{census.demographics.list}
\alias{census.demographics.list}
\docType{data}
\title{
census.demographics.list
}
\description{
A list of the demographic variables in the SF1 (\url{http://www.census.gov/prod/cen2000/doc/sf1.pdf}) tech report. Primarily for use in \code{\link{demographics.add}}.
}
\usage{data(census.demographics.list)}
\format{
List of characters.
}
%\details{}
%\source{}
\references{
Census 2000 Summary File 1 [name of state1 or United States]/prepared by the U.S. Census 
Bureau, 2001. \cr
\url{http://www.census.gov/prod/cen2000/doc/sf1.pdf}
}
\examples{
data(census.demographics.list)
}
\keyword{datasets}
