\name{state.ab}

\alias{state.ab}
\docType{data}
\title{
State Abbreviations and DC for all states available in UScensus2000-suite.
}
\description{
For use in \code{\link{demographics.add}}..
}
\usage{data(state.ab)}
\format{
Vector of characters.
}
%\details{}
%\source{}
%\references{
%%  ~~ possibly secondary sources and usages ~~}
\examples{
data(state.ab)

}
\keyword{datasets}
